Add files via upload
1.	XY_train.csv in folder ‘’Task 1’’ contains a dataset of Views against 6 input variables.

a.	Select and train a model to predict the views for a given video, given the features provided.
b.	Explain your choice of model and why it is suitable for this problem
c.	Comment on your model’s performance as well as the significance of each feature.
d.	Provide predictions for views of videos within X_test.csv, as well as an estimation of your out-of-sample performance.
1.	XY_train.csv in folder ‘’Task 1’’ contains a dataset of Views against 6 input variables.

a.	Select and train a model to predict the views for a given video, given the features provided.
b.	Explain your choice of model and why it is suitable for this problem
c.	Comment on your model’s performance as well as the significance of each feature.
d.	Provide predictions for views of videos within X_test.csv, as well as an estimation of your out-of-sample performance.